feat: consolidate profile, engagements and achievements into one page (#505)
Replaces the three separate pages (Profile, My Engagements, Achievements)
with a single ProfileOverviewPage at /profile, using a tab bar to keep
each section clearly distinct. Old routes /my-engagements and /achievements
redirect to the unified page with the appropriate tab selected via ?tab=.
Header dropdown is simplified to a single "My Profile" link.
