+++ frontend/src/components/Header.tsx
@@ -302,25 +302,6 @@ export default function Header() {
 											</div>
 											<div className="py-1">
 												<Link
-													to="/my-engagements"
-													className={`flex items-center gap-3 px-4 py-2.5 text-sm transition-colors ${isTransparent ? "text-white/80 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-700"}`}
-												>
-													<svg
-														className="w-4 h-4"
-														fill="none"
-														viewBox="0 0 24 24"
-														strokeWidth="1.5"
-														stroke="currentColor"
-													>
-														<path
-															strokeLinecap="round"
-															strokeLinejoin="round"
-															d="M9 12.75 11.25 15 15 9.75M21 12a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z"
-														/>
-													</svg>
-													{t("nav.myEngagements")}
-												</Link>
-												<Link
 													to="/profile"
 													className={`flex items-center gap-3 px-4 py-2.5 text-sm transition-colors ${isTransparent ? "text-white/80 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-700"}`}
 												>
@@ -338,25 +319,6 @@ export default function Header() {
 														/>
 													</svg>
 													{t("nav.myProfile")}
-												</Link>
-												<Link
-													to="/achievements"
-													className={`flex items-center gap-3 px-4 py-2.5 text-sm transition-colors ${isTransparent ? "text-white/80 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-700"}`}
-												>
-													<svg
-														className="w-4 h-4"
-														fill="none"
-														viewBox="0 0 24 24"
-														strokeWidth="1.5"
-														stroke="currentColor"
-													>
-														<path
-															strokeLinecap="round"
-															strokeLinejoin="round"
-															d="M16.5 18.75h-9m9 0a3 3 0 0 1 3 3h-15a3 3 0 0 1 3-3m9 0v-3.375c0-.621-.503-1.125-1.125-1.125h-.871M7.5 18.75v-3.375c0-.621.504-1.125 1.125-1.125h.872m5.007 0H9.497m5.007 0a7.454 7.454 0 0 1-.982-3.172M9.497 14.25a7.454 7.454 0 0 0 .981-3.172M5.25 4.236c-.982.143-1.954.317-2.916.52A6.003 6.003 0 0 0 7.73 9.728M5.25 4.236V4.5c0 2.108.966 3.99 2.48 5.228M5.25 4.236V2.721C7.456 2.41 9.71 2.25 12 2.25c2.291 0 4.545.16 6.75.47v1.516M7.73 9.728a6.726 6.726 0 0 0 2.748 1.35m8.272-6.842V4.5c0 2.108-.966 3.99-2.48 5.228m2.48-5.492a46.32 46.32 0 0 1 2.916.52 6.003 6.003 0 0 1-5.395 4.972m0 0a6.726 6.726 0 0 1-2.749 1.35m0 0a6.772 6.772 0 0 1-3.044 0"
-														/>
-													</svg>
-													{t("nav.myAchievements")}
 												</Link>
 												<button
 													type="button"
@@ -607,25 +569,11 @@ export default function Header() {
 									<OrganizationSwitcher transparent={isTransparent} />
 								</div>
 								<Link
-									to="/my-engagements"
-									onClick={() => setMobileOpen(false)}
-									className={`block px-3 py-2 rounded-lg text-sm font-medium transition-colors ${isTransparent ? "text-white/90 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-600"}`}
-								>
-									{t("nav.myEngagements")}
-								</Link>
-								<Link
 									to="/profile"
 									onClick={() => setMobileOpen(false)}
 									className={`block px-3 py-2 rounded-lg text-sm font-medium transition-colors ${isTransparent ? "text-white/90 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-600"}`}
 								>
 									{t("nav.myProfile")}
-								</Link>
-								<Link
-									to="/achievements"
-									onClick={() => setMobileOpen(false)}
-									className={`block px-3 py-2 rounded-lg text-sm font-medium transition-colors ${isTransparent ? "text-white/90 hover:bg-white/10 hover:text-white" : "text-gray-700 hover:bg-brand-50 hover:text-brand-600"}`}
-								>
-									{t("nav.myAchievements")}
 								</Link>
 								<button
 									type="button"
