--- frontend/src/pages/AccountPage.tsx
@@ -1,5 +1,0 @@
-import { Navigate } from "react-router";
-
-export default function AccountPage() {
-	return <Navigate to="/profile" replace />;
-}
